Move copied code to it's own file to avoid code duplication and import loop
